requirements.txt
requests==2.8.1
PyYAML==3.11